Remove ArtworkTerm model and artwork include
I believe this is a remnant from a time when we would've had to import
terms by hitting a separate endpoint upstream. This was due to a mis-
understanding, wherein it was thought that term type – style, material,
classification, etc. – was intrinsic to the artwork-term relationship,
rather than to the term itself. Once that misunderstanding was cleared,
we now receive terms as two arrays on each artwork (pref and alt) and
generate the outbound fields on request, based on the hardcoded type.