Remove resource2gallery method from TicketedEvent
We've been informed that in the source system, content editors reuse
resources for different events – the ids might stay static, but which
locations they represent can change over time. As such, this isn't a
reliable way of connecting events to locations, esp. not hard-coded.
With the new website, we might be able to assign locations to events
in a normalized manner.