Refactor Fillable into inbound transformers
Massive commit. Essentially, inbound transformers are meant to be the
glue between import commands (or the `ImportsData` trait) and Eloquent
models. Previously, this logic lived in the models themselves, and in
the `Fillable` trait. Moving it to dedicated classes gives us a place
to refactor, clean, and track it better.

I'm going to take a minute to explain how it works.

First, one of the import commands (e.g. `import:collections`) calls the
`import` method from the `ImportsData` trait. In that call, it sets the
source system to "Collections" and specifies which model class is being
imported.

The `import` method then calls `getInboundTransformerForModel` defined
in `ResourceServiceProvider`, which returns the inbound transformer
class associated with the given model and source system combination.

Currently, this mapping is generated via naming conventions, but later,
we can create a manual mapping, e.g. in `config/resources.php`. As-is,
we can only have one mapping per model, but that's easy to change.

The `import` method then calls `save` (cf. `AbstractImportCommand`),
passing it the model and transformer classes, as well as the object
gathered from the source system.

Inside the `save` method, we create a new instance of the inbound
transformer, passing the raw object and model to its methods.

In the inbound transformers, we turn the raw object into an instance
of the new `Datum` class. This class brings with it some conveniences,
e.g. empty strings are treated as null, referencing non-existent props
returns `null` instead of throwing exceptions, etc. This is great for
creating consistency across models.

The inbound transformers are architected with two primary principles:

 1. Use inheritence to stack transformations
 2. Avoid calling model methods whenever possible

Generally speaking, most methods should return arrays, without making
any modifications to the models. The arrays get merged along the way,
and the methods in `AbstractTransformer` use them as arguments for
calling `fill` and `sync` on the model. This is a major change in
design from before: we only sync once, we only filter attributes
once, etc. There is room for hard-coding relations (cf. `syncEx`),
but generally, we should avoid doing so whenever possible.

Alright, that's how it works, but what else has changed?

 1. We no longer import "pivot" resources directly. All pivots are
    still imported, but through their primary model, not separately.

    This functionality depends on the following commits:

    https://github.com/art-institute-of-chicago/data-service-collections/commit/b0362c6
    https://github.com/art-institute-of-chicago/data-service-collections/commit/167ff5c

 2. As a side-effect of above, the following need more work:

 	- Agent's placePivots
 	- Artwork's cataloguePivots
 	- Exhibition's venuePivots

 3. In many cases, things should detach much better now. For example,
    if an artwork went from having three publish categories to having
    zero within one update, that change wouldn't have been registered,
    because `$source->category_ids` would be falsey. Now, cases like
    this should cause related items to detach properly everywhere.

 4. Gallery's `type` is no longer updated during artwork import.
    This is a remnant from when Galleries were Places.

That's about all. This needs a sweep for more bugfixes, and the pivot
issue really needs to be resolved for consistency. The manual-mapping
question should be revisited later.