Fix and improve imports for selections
There's a few things here. First, `content` has been renamed to `copy`
on `selections`. This allows us to use the `HasBlocks` trait to extract
text copy from the blocks within. We could work around the field name
disparity, but it's nice to have this consistency.

Consequently, the `HasBlocks` trait has been expanded to extract text
content from non-paragraph blocks. It's not magic - we'll need to add
mapping rules to the `getCopyRules` method, but it's a nice start.