Adjust limit in `query` override
There's an issue w/ `agents` due to agent places. There's too many
places, and the server chokes trying to send too many queries. The
solution would be to implement eager loading, but I'm not sure that
it makes sense to do so before migrating the collections data service
from Ruby+Grape to Laravel. For now, we'll just lower the limit.