Hotfix `touch` for `BelongsToMany` relations
This makes `$touches` work with `BelongsToMany` relations. Before, it
only really worked for `BelongsTo` relations in our use-case.

We're trying to make it so that Scout's `searchable` gets called when
a model gets touched by a related model that just got updated.

This solution is not ideal for two reasons: it can result in infinite
loops, and it causes `saved` events to fire for each instance in each
relation in `$touches`. For our use-case, we should back off and try
to batch-update `touch` and/or `searchable` calls.