Hotfix M2M departments for artworks
We thought that artworks would only ever be assigned a single
departmental category. That's incorrect, they can be assigned
multiple, but only one is preferred, but there's no good way
to set which is preferred upstream, so we need to reference
category hierarchy. Anyway, we can resolve that later, this
is an acceptable solution for the moment.