Allow POST to retrieve data from REST API
This commit is being made (1) to increase parity b/w the search and
REST APIs, and (2) b/c our website's codebase can only set the read
method globally, not per-call. It needs to use POST everywhere safely.

Using `Route::any` is safe, because it doesn't override OPTIONS:

https://github.com/laravel/framework/issues/14858

Foundation was updated to allow `POST` but throw an exception on other
methods. It'd be better to give feedback to the user via the exception
handler, but this'll do for now.