Refactor `pageviews` to use `function_score`
Precursor commit to implementing `boost_rank`, which will allow us to
semi-manually control the order of the top-n artworks. This creates a
space for that logic to live.

https://www.elastic.co/guide/en/elasticsearch/reference/6.0/query-dsl-function-score-query.html

This does affect the scoring of artworks - the logic isn't completely
equivallent to what it was before - but in a good way! Running test:

   phpunit tests/Functional/ArtworkSearchTest.php

...shows that we now fail 6 tests, instead of 9.