Remove `/boosted` endpoints
These endpoints no longer accurately represent our boosting logic. They
only show entities with `is_boosted` true, so they don't capture the
`boost_rank` logic, nor the `pageviews` logic. As such, they are more
misleading then helpful. Users should be encouraged to hit `/search`
with `artworks` specified as a resource instead.

(Additionally, `/boosted` doesn't show the same order as `/search`.)

It could be argued that the `/boosted` endpoints should just be changed
to show entities that are boosted in any way whatsoever, but given that
one of our long-term goals is to pull in `pageviews` for ALL artworks,
this would not be a transparent approach.

We need to regen the docs, but `/boosted` is gone from Documentable.

Mysteriously, `/boosted` tests weren't failing on my local box, but
they've been removed nonetheless.