Remove `/_search` endpoints
Just to be clear – the `/search` endpoints still exist! None of our
services (including data-dashboard) should be affected by this change.

The only reason we included the `/_search` endpoints initially was to
emulate Elasticsearch conventions. We hoped that users of our API might
be able to use official Elasticsearch clients to query it. However, as
the project evolved, this became both impossible and undesirable from
our perspective. As such, there's no reason to keep these endpoints.