Add `style_title` and `classification_title` to Artwork
This is required to support "Explore Further" functionality on the new
website. Specifically, this removes the need to perform additional
queries to retrieve titles for preferred terms based on id to display
tab labels. For that purposes, `artist_title` is also required, but
it's already available in the API.

Still hoping to move to a more nested structure for `/api/v2`.

Requires reindex, but since we are relying on Elasticsearch's defaults
for string fields (text w/ keyword subfield), we can run it softly.