Transform non-HTML artwork descriptions to HTML
This uses a Markdown parser in the CommonMark syntax to convert
plaintext, non-HTML or mixed-HTML text to HTML. This allows CITI
users to stop worrying about inputting things like paragraph tags
manually, while still ensuring that the content gets rendered well
by applications that expect HTML, and preserving custom markup such
as `<em/>` and `<strong/>`.

The one diversion made here from CommonMark is to treat soft breaks
as hard breaks. Per spec, if you want a hard break, you need to add
two spaces at the end of a line before a break. This is much maligned
for many reasons - it has its uses, but it's not intuitive for non-
technical users. For this reason, we'll be using the league parser's
built-in `soft_break` override.

https://stackoverflow.com/questions/1946426/html-5-is-it-br-br-or-br

Run `php artisan update:descriptions` to update existing data.