Remove unused TicketEventType inbound transformer
This is a left-over from a refactor in `data-service-events`:

https://github.com/art-institute-of-chicago/data-service-events/commit/a216765

That commit forces multiple over-writes on the same model instances,
ensuring there's no duplicate event types. The de-dupe logic is in
the data service now, not here.

By removing this transformer, TicketedEventType will fallback to the
MembershipTransformer during import (see `getInboundTransformerMapping`
in `ResourceServiceProvider`), which binds `membership_id` to id.

This is also an interesting case in that the filename did not match
the classname. If the `import:events-ticketed-types-full` was run
directly after running `git pull` without running `composer install`,
it wouldn't create the file-to-class-map, so the transformer wouldn't
be registered; it'd fallback to PSR-4 autoloading, but because the file
was misnamed, it'd fail to find it, fallback to MembershipTransformer,
and the import command would succeed. But you ran `composer install`,
it'd register this class for autoloading, and the command would fail.

The specific error was in MySQL:

  Integrity constraint violation: 1048 Column 'membership_id' cannot be null

Anyway, removing this file fixed these issues.