Remove `in_on_view` boost for Artwork
We are moving towards making pageviews the only adjustment to relevancy
scoring. This specific boost introduced an issue w/ artworks that the
institution has taken off-view for external loan or conservation. For
example, American Gothic was off-view because it was being exhibited
elsewhere. However, this didn't make it any less popular, or less
significant, so people were surprised that it wasn't popping up
higher in their search results.