Refactor addFunctionScore to use search scopes
This method is meant to run as the last step before sending the query.
It'll take the query generated thus far, and replicate it, such that
each resource with a `function_score` query gets its own copy, within
a search scope targeting that resource. All other resources without
`function_score` should also end up with a copy of the query inside
a scope that targets just them.