Refactor `touchesOwner` method to execute a reindex, rather than update the database timestamp.
This is a minor refactor that will have some fairly widespread implications.

The reasoning behind this refactor is that we're seeing some major slowdowns in import processes from our source systems when records are associated with thousands of other records. When we have multiple imports like this at a time, the problem compounds and our processes eventually choke.

Part of why we first implemented the `touches` logic was because related elements were staying stagnant in the search index. This refactor skips the records-are-reindex-when-they're-saved functionality and calls for the related records to be reindexed directly. The downside is that timestamps of related models will not be updated (perhaps this logic was providing red herrings and causing other issues?). The upside is that this new refactor is no longer reliant on database transactions, and can ping a list of records to be reindexed much more quickly. Once we move back to Scout running on queues, we won't have to wait for these processes to finish before our imports are "complete."