Dramatically boost phrase clause [WEB-22]
We want to ensure that when a user issues multi-keyword searches,
we prioritize docs that have all of their words in proximity, and
ideally, in the same order.

This improves searches for artwork names, artist names, and some
tagging terms. A litmus test we've been using is searching for
"African American" and checking that "African" or "American"
artworks don't get returned higher in the results than works
which mention the whole phrase.