Add api/v1/event-occurrences for mobile [WEB-511]
Most of the weird stuff here is happening on the website end, not here.

Note that the endpoint is defined in the command, but `importFromWeb`
doesn't actually get called, so it won't be imported as part of running
`import:web-full` – you need to explicitly specify the endpoint:

    php artisan import:web-full events/occurrences

This is because `import:web` inherits from `import:web-full`, and the
EventOccurrence model is auto-incrementing to avoid id conflicts,
because in the website's API, all event occurrences share their
master event's id, but we don't want running `import:web` to
create record duplication.