Make all timestamps null by default
Before, some timestamp columns used to be `DEFAULT CURRENT_TIMESTAMP`.
This is bad from a data integrity standpoint. It's only useful if we
have a single timestamp column indicating when a record was added to
the database. Even then, we want the timestamp of when a record
was imported into the app, not when it was added to a database,
which can differ, e.g. if we use queues.