Remove dates from pivots and CITI lists
This one is complex to explain, but in short, some of our collection
resources, most notably terms and categories, are stored as "lists"
in CITI, our collection management system. CITI does not track the
last modified time for each list item. However, our DAMS, LAKE,
which currently serves as an intermediary between CITI and the
data hub, ascribes false last modified times to these items on
each import. Because of that, we think that each of these items
have been modified, when in fact, they have not. We cannot trust
the last mod times on these items. We need to check their contents
to determine whether or not we should import them.

Importing them needlessly bogs down our system, esp. when you
consider that each item touches artworks that reference it.

In regards to pivots, CITI does not provide last mod times for them
either. This has no impact on performace on our end, but it does make
tracking that information unnecessary.