Move analytics and ULAN to import:monthly
Doing this to speed up incremental imports. It makes sense to move
these two items to daily down the line, but it's not necessary at
the moment. The ULAN matching is most directly useful for matching
collections records to library data. The analytics will be useful
for when we are prioritizing recently popular artworks, but for
now, it's too stable for one month to make much of a difference.