Remove PullRecord webhook
This is a temporary measure: I need to restructure the way sources
are defined in the resource config, and I'd like to clean up the
ImportsData trait, but having this functionality made that tricky.

The webhook was requested by our website vendor, but it ended up
not being used for launch. I'd love to reimplement it, but I need
to clean up a few things first. Note that the structure expected
by the pull record endpoint will almost certainly change.