Set foundation requirement to rolling version
This follows the convention established by our website, where we always
require the most recent version of `aic/data-hub-foundation`, but we
are very selective about when we update it. Before, we needed to tag
just about every commit in the library in order to update dependents,
which felt strange.

The new approach requires us to be very selective about what we update.
However, this should already be the case. Never run this:

    composer update

This updates your entire dependency tree, which is quite likely to
cause things to break somewhere in a way that is difficult to debug,
despite everyone's best attempts at semantic versioning. Instead,
try to update one package at a time:

    composer update aic/data-hub-foundation

You can also update a package and its dependency tree:

    composer update aic/data-hub-foundation --with-dependencies

Note that this won't update any dependencies that are also direct
dependencies of your "root" project. This approach can still result
in problems, but it's easier to fine-tune than updating all
dependencies at the same time.