Force `lf` line endings on git checkout
My reasoning is that this software stack is 100% not designed to run
on Windows. I'm currently coding on Windows, and ultimately, my text
editor, terminal, and git tooling handle `lf` just fine. Checking out
as `crlf` creates issues when files need to be shared between a host
and a virtual machine (e.g. Homestead). Might as well just force git
to always check out as `lf`. Future users, feel free to open an issue
if this is a bad idea!