Filter fields before evaluating mapping values [WEB-114]
This makes it so that passing `fields` params to (non-search) REST API
endpoints is now much more efficient. Previously, we would filter the
wanted fields *after* we already evaluated `value` for all fields.

There's still two outstanding issues. First, we eager load all related
models for artworks for every request, even if the fields they are used
to fill weren't requested [WEB-115]. Secondly, we still retrieve all
fields from the data base. There's a possibility that we could use the
`fields` param to inform which columns to select.

The first issue should be relatively easy to fix, but the second will
be much harder, as it'll require a reverse mapping, wherein each API
field declares which database fields it needs to render itself.