Remove all subresource routes
I feel bad doing this, but I think it has to be done at least for the
time being. The biggest factor in this decision is that this pattern
was not consistently implemented across the codebase. If we were to
offer subresources as a feature of our API, it should be implemented
for *all* relationships, on *all* resources. We have dozens of such
relationships, but only seven subresource routes.

Given that our API might be public soon, it makes sense to address
as many inconsistencies as possible prior to making it so. None of
our internal apps depend on subresource routes, so now is the safest
time to remove them if that is to be our decision.

The alternative would be to implement subresources for *all* resources.
However, given the current architecture, doing so would require a lot
of "hardcoding" `forFoobar` methods, custom tests, and so on. It's just
not a sustainable way of going about things, as it creates yet another
thing we need to remember to manually upkeep.

I think we'll have a relatively easy way to reimplement this if we so
choose down the line, after we improve relationship mapping. Removing
subresource routes will make that refactor easier.

Lastly, I'm not sure if subresources are an API feature that we should
support down the line. I feel like the same functionality can be done
through search - granted, it's more verbose, but also more flexibile.
I think that consolidating functionality and offering a single way to
do something will generally speaking improve performance (caching) and
reduce overhead for us in relation to upkeep.

However, this is very much an arguable opinion, not necessarily best
practice. Not to bury the lead: putting aside the question of whether
or not we should support subresources in the future, it still makes
sense to remove them now, given the points raised above.