Replace ArtworkController with ConfigController
We are now in a place where we can get rid of our model-specific
controllers entirely. We've reduced them to settings: none of them
really do much more than define `model` and `transformer` pairs.
We can now move these "outbound" settings to live with our other
resource definitions in `config/resources.php`.

The benefits of doing so are twofold.

First, it reduces the overhead of adding new resources to the
codebase. Doing so already requires defining the resource in the
config, so adding another line is less overhead than having to
also create new classes.

Secondly, it opens the door to being able to more easily define new
model-transformer pairs for future API versions. I think this will
become important once our API is public. Again, this claim rests on
the premise that adding config is cleaner than creating new classes.

If we ever need to create a controller with truly custom functionality,
it will be easy to create a custom route that binds to that controller,
rather than to `ConfigController`. This commit shows how both types of
routes can coexist.