Use config/resources.php to generate routes/api.php
The config is now the SSOT for routes. When creating new resources, it's
no longer necessary to add a resource definition to the config *and* add
a route. It's a bit magic, but with documentation, it should make the
process of adding new resources easier.

Looking forward, I think following this pattern may ease API versioning,
if we decide to go that route.