Remove `exampleId` declarations from models
This is a useful function, but it is unfortunately fragile. In many
cases, we've seen our example objects disappear from upstream systems
between releases, causing doc generation to break. There are some cases
where we can assume them to be near-permanent (e.g. Nighthawks), but in
most others, this is not the case.

I think this also causes some confusion over how we are supposed to be
treating seeding in our database. Like, I think we can safely assume
that the data-aggregator as-is is meant specifically for AIC's systems.
We never meant to make it plug-and-play for other institutions. From
that perspective, we can assume that if there is real data in the
database, it's from AIC. The old `exampleId` method reflected this
approach, in that it assumed that the database was indeed populated
with live data from AIC.

However, at least at the beginning of the project, we started with the
idea that a user should be able to download this repository, seed the
database with fake data, and pretty much operate it normally in terms
of output. I feel like documentation falls into "output" territory.
But the old `exampleId` implementation was incompatible with a seeded
database! This new implementation should make it compatible.

That said, stepping back for a moment, I'm not sure that we should be
pushing potential contributors towards seeding, nor investing time in
building out seeding capabilities any more than is needed to support
testing. I see the value in it, but I think a better investment would
be to build out infrastructure around database dumps.

In lieu of seeding, we could dump the database on a periodic basis into
CSV files, offer these dumps via GitHub, and write a command to import
these dumps into the data-aggregator. This should get collaborators
ready to contribute quicker and in a more rewarding way than if
they had to warm up with seeded databases.

That said, any model can still define its own `exampleId` to override
this new default behavior.