Add Documentable trait to all models
I think we used to only apply this trait to models that should make it
into the docs. However, I think this resulted in at least one model
with an endpoint being omitted (AgentType). That's a pretty good track
record, especially given that we haven't generated docs in a while.

Still, it can get tedious to track where Documentable should and should
not be applied, especially in terms of inheritance. With that in mind,
I don't think it hurts anything to just apply it to all models. It's
just some additional methods, it shouldn't change anything intrinsic
about the model. For models that don't use it, it won't affect
anything, and it's one less thing for us to track.

In terms of excluding models from documentation, we can easily track
that data in `config/resources.php`. I think at one time, we might've
been checking if a model has the Documentable trait, but that has
proved unreliable. I'd like to gradually move towards thinking about
generating documentation "for endpoints" rather than "for models".
I think that's a requirement if we ever decide to do API versioning.