Further reduce Scout chunk size
Amazingly, halving the chunks wasn't enough. I've reduced it to 1/5th
of the original value, and it seems to be going through.

Once we introduce some sort of job system, it should be relatively easy
to check if the index request fails, and schedule recovery jobs which
halve the previous batch size.