Throw exception on Elasticsearch update error
This isn't ideal because we use the bulk update API, so some of the
bulk items can succeed, while other fail. In an ideal situation, we
would requeue only the items that failed, and allow the bulk process
to continue. We aren't quite there yet, so this is a good stopgap.
We are ready to survey these exceptions via Sentry.