Remove `latitude` and `longitude` from API for places
We still provide these for galleries. I'm sad to remove these, but we
were told that no one has audited the coordinates for non-gallery places
for a long time, and the institution considers them unreliable. Once we
are receiving normalized data for places, we should reconcile them and
import new data from an external source. We might be able to use the
existing coordinates to aid in the reconciliation.