Remove Category relationship from Gallery and Place
This might have been something that was provided in the old website, but
it seems that it is not being provided any longer. There are no records
in that table, there's nothing in the designs that expects this data,
and the upcoming CITI API changes make no mention of fields that would
provide it.