Add Web\Location transformer
Another resource we should consider deleting. There's only two. I think
initially, we were thinking this would be used for events. But it's just
the two entrances, and content editors already have enough forms to fill
out when scheduling events that I don't think locations would be doable
in the near-term. Just need to check if the CMS depends on our search.