Refactor search-related fields into transformers
All fields are now visible both in search and in the REST API. Should
ease debugging. The process of defining suggest fields could be cleaner.

Note the new `filter` key in field definitions. If you need to ensure
that a field gets left out entirely, rather than just being `null`, you
can pass a function with a structure similar to that of `value`, which
returns a boolean. On `false`, the field will be omitted.