Remove suggestions from search endpoint
I'm hesitant to make this change. I think it's worthwhile doing, since
for users of the data hub, typically, getting suggestions is a separate
step from actually executing the search. Suggestions are meant to be
delivered via fast, light endpoints, not mixed in with search.

Tentatively, if we implement "Did you mean?"-style suggest results on
the website, it might make sense to revert this change. However, we do
not offer that functionality currently, and there's no plans to offer
it in the future. I'd like to remove this in preparation for opening the
API, since we don't want API users to become dependent on functionality
that hasn't been audited by internal use.

Additionally, this functionality only delivers suggestions from the
autocomplete endpoint, though we want to push users towards autosuggest.
We could implement both, but we should consider doing some refactoring
in the search client first.