Hide unused mobile fields in artworks [WEB-79]
We are about to Cantor pair TourStop id's which causes out-of-bounds
errors for `tour_stop_ids` on this endpoint. I decided to do this after
examining the mobile apps' codebases. Since these fields aren't being
used by the app currently, there's no oversight to ensure they remain
bug-free, and it's good to slim things down for the public release.

If this is a poor decision, let's discuss and revert!