First pass at getting authentication set up [WEB-133]
I'm currently stuck on the Frontend Quickstart section of the Passport
documentation. We've never used Laravel's built-in frontend before, so
much of the work in this commit has been re-engineering that to get it
to function at a basic level. This commit represents all the basic
pieces in place, with added Vue components for Passport, but the compiled
frontend doesn't seem to recognize the new component tags.