Remove `assets` alias in Elasticsearch [WEB-95, WEB-848]
We ran into an issue in Elasticsearch 6.0 wherein we could not target
`content.keyword` subfield because it only existed (at the time) in the
`videos` index, not in all the other indexes under the `assets` alias.
Once we switched to specifying the indexes directly, rather than using
`assets` in our query, this issue disappeared. It's possible that this
bug was fixed in a more recent version of Elasticsearch, but we want to
discourage the use of aliases to avoid weird bugs like this one. All of
our properties target individual indexes. Plus, the presence of `assets`
is somewhat of an artifact of the fact that these resources are all
stored in one system (our DAMS).