Run `braces` to fix methods and functions [WEB-759]
Keeping `braces` disabled because it made many other changes. I think
it can be useful to run it occasionally to help identify things to fix
manually, but I don't want it getting in our way. It also bugs out in
a few situations, such as when comments immediately follow an opening
brace. I know it's meant to adhere to PSR-2, but the changes it makes
are a bit too widespread to accept all at once.