Switch from HelperServiceProvider to composer autoload [WEB-1197]
This is how Laravel does it in vendor\laravel\framework\composer.json:

    "autoload": {
        "files": [
            "src/Illuminate/Foundation/helpers.php",
            "src/Illuminate/Support/helpers.php"
        ],
        "psr-4": {
            "Illuminate\\": "src/Illuminate/"
        }
    },

This makes the helper functions available before the framework loads.
PHPStan likes this. Without it, it complained about stuff like this:

    ------ ------------------------------
     Line   Http/Search/Request.php
    ------ ------------------------------
     897    Function hexToHsl not found.
    ------ ------------------------------

If we don't like this approach, we can either specify these files
to be autoloaded manually in `phpstan.neon.dist` as described here:

https://github.com/phpstan/phpstan#autoloading

...or we can turn these two files into classes similar to Larave's
`Arr` and `String`, forgoing global functions.

Now we just need to make it recognize Laravel facades...