Fix all Level 0 errors for PHPStan [WEB-1197]
PHPStan is having lots of trouble with Laravel "magic" so to speak. In a
separate branch, I tried using this instead:

https://github.com/nunomaduro/larastan

...but it still didn't properly handle most of the cases that ended up
getting added to `ignoreErrors` in `phpstan.neon.dist`

It's possible to write custom rules to handle this "magic", but it's too
early for that. Let's just do a sweep, ignoring everything specific to
Laravel for now.