Remove `send` fields from email series pivots [WEB-1236]
The checkboxes in the CMS are used to determine if the pivot should
override default copy, not to choose if the email should be sent to
specific audiences. The event editors have no choice over the choice
of audience here: that is determined by the choice of email list and
by the related event host information.