Remove unneeded override of VerifyCsrfToken
Our "fix" to remove CSRF token verification only worked because we
were removing the requirement to execute this verification on all posts.
It appears the problem was due to the forms being cached, and passing
stale tokens to the controller. We cleaned up the CDN to not cache the
forms, so CSRF verification now works.