Remove manual word wrap from readme
OTOH, I respect that hard line breaks slash manual word-wrap is good
practice, particularly when reading docs on the commandine. OTOH, MD
files are typically word-wrapped at 80 chars per line, while this file
was wrapped around 120 chars or so. It's not that big of a deal, but
I think if we are manually wrapping docs, we should follow recommended
expectations. But also, editing manually wrapped files is a bit of a
pain. I think there are editor plugins that will re-wrap files for you,
but I'm not sure to what extent they account for URLs and similar edge
cases. For expediency, I'm removing the wrap.