Disable `it_403s_if_limit_is_too_high` [WEB-1382, WEB-1189]
We made it so that authenticated users are not subject to the limit max.
This also kicks in if `APP_RESTRICTED` is false, and right now, it's
always false during testing. As a result, this test always returns 200.
Removing this test until we can get a baseline of what's working...