Remove hours from API [WEB-1261, WEB-1194]
At one time, we thought that hours might be a more complex system than
it ended up being. Our actual hours on the website are a free-text
field, more of a scheduled alert than something programmatic in the
event-planning sense. And over time, our API diverged from how this
data is actually represented in the source system, because we don't
really dogfood this endpoint. As such, it makes sense to just delete
it until it's actually needed for some reason, by someone who will
provide ongoing oversight.