Refactor unit tests to run on SQLite [WEB-1801]
We used to run unit tests on a local MySQL database. For the sake of speed, we refactored the tests to run on a local SQLite database. We have tests running on a flat-file database rather than an in-memory database so that the migrations persist from test to test. With an in-memory database, all migrations get rerun before every test.